Create README.md
Password Manager.
It is a graphical user interface application that can be used to store the password 
along with the username. All the data will be encrypted to prove better security 
and will be decrypted on demand. Every password will also be hidden behind a 
wall which we can enter by using the application password.
Technologies Required: C++ programming language, GUI library,
Encryption/Decryption, Data Structure, and Algorithms, Secured Storage.